refactor: streamline component imports and update type definitions for consistency
