+++ src/App.tsx
@@ -1,9 +1,5 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Layout from "./components/common/Layout.tsx";
-import Homepage from "./pages/Homepage.tsx";
-import CartPage from "./pages/Cartpage.tsx";
-import Detailpage from "./pages/Detailpage.tsx";
-import Loginpage from "./pages/Loginpage.tsx";
+import { Layout, Homepage, Cartpage, Detailpage, Loginpage } from "./utils";
 
 function App() {
   return (
@@ -12,7 +8,7 @@ function App() {
         <Routes>
           <Route path="/" element={<Homepage />} />
           <Route path="/shop" element={<Homepage />} />
-          <Route path="/cart" element={<CartPage />} />
+          <Route path="/cart" element={<Cartpage />} />
           <Route path="/item/:name" element={<Detailpage />} />
           <Route path="/login" element={<Loginpage />} />
         </Routes>
