+++ src/components/cart/index.ts
@@ -1,0 +1,1 @@
+export { default as Cart } from "./Cart.tsx";
