+++ src/components/common/index.ts
@@ -1,0 +1,14 @@
+// Desc: Export all common components
+export { default as Header } from "./Header.tsx";
+export { default as Footer } from "./Footer.tsx";
+export { default as Img } from "./Image.tsx";
+export { default as Heading } from "./Headingtxt.tsx";
+export { default as Input } from "./Input.tsx";
+export { default as Button } from "./Button.tsx";
+export { default as LoadingSkeleton } from "./LoadingSkeleton.tsx";
+export { default as UspBanner } from "./UspBanner.tsx";
+export { default as PageBanner } from "./PageBanner.tsx";
+export { default as Pagination } from "./Pagination.tsx";
+export { default as Breadcrumb } from "./Breadcrumb.tsx";
+export { default as Loading } from "./Loading.tsx";
+export { default as Layout } from "./Layout.tsx";
