+++ src/components/detailpage/ProductDisplay.tsx
@@ -2,10 +2,10 @@ import { useState } from "react";
 import styles from "../../styles/components/detailpage/ProductDisplay.module.scss";
 import Button from "../common/Button.tsx";
 
-interface ProductDisplayProps {
+type ProductDisplayProps = {
   images: string[];
   name: string;
-}
+};
 
 const ProductDisplay = ({ images, name }: ProductDisplayProps) => {
   const [selectedImageIndex, setSelectedImageIndex] = useState(0);
