+++ src/components/detailpage/index.ts
@@ -1,0 +1,3 @@
+// export Homepage Components
+export { default as Related } from "./ProductRelated.tsx";
+export { default as ProductInfo } from "./ProductInfo.tsx";
