+++ src/components/filters/index.ts
@@ -1,0 +1,1 @@
+export { default as FilterSortBar } from "./FilterSortBar.tsx";
