+++ src/components/home/index.ts
@@ -1,0 +1,2 @@
+export { default as ProductCard } from "./ProductCard.tsx";
+export { default as ProductGrid } from "./ProductGrid.tsx";
