+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export { default as Homepage } from "./Homepage.tsx";
+export { default as Cartpage } from "./Cartpage.tsx";
+export { default as Detailpage } from "./Detailpage.tsx";
+export { default as Loginpage } from "./Loginpage.tsx";
