+++ src/utils/index.ts
@@ -1,19 +1,21 @@
-// Exporting all the components for easier importing
+// common Components
+export * from "../components/common";
 
-export { default as Header } from "../components/common/Header.tsx";
-export { default as Footer } from "../components/common/Footer.tsx";
-export { default as ProductCard } from "../components/home/ProductCard.tsx";
-export { default as Img } from "../components/common/Image.tsx";
-export { default as Heading } from "../components/common/Headingtxt.tsx";
-export { default as Input } from "../components/common/Input.tsx";
-export { default as ProductGrid } from "../components/home/ProductGrid.tsx";
-export { default as Button } from "../components/common/Button.tsx";
-export { default as LoadingSkeleton } from "../components/common/LoadingSkeleton.tsx";
-export { default as UspBanner } from "../components/common/UspBanner.tsx";
-export { default as PageBanner } from "../components/common/PageBanner.tsx";
-export { default as FilterSortBar } from "../components/filters/FilterSortBar.tsx";
-export { default as Pagination } from "../components/common/Pagination.tsx";
-export { default as Breadcrumb } from "../components/common/Breadcrumb.tsx";
-export { default as Icons } from "./Icons.tsx";
-export { default as Related } from "../components/detailpage/ProductRelated.tsx";
+// Layout
+export * from "../pages";
+
+// Cart
+export * from "../components/cart";
+
+// Home
+export * from "../components/home";
+
+// filters
+export * from "../components/filters";
 export * from "./helperFunctions.ts";
+
+// detailpage
+export * from "../components/detailpage";
+
+// Icons
+export { default as Icons } from "./Icons.tsx";
